refactor: replace batch upload with single-sync sessions
Assets were mutated through a multi-batch upload protocol where the
plugin created a batch, uploaded chunks, then committed operations. The
canister supported many concurrent batches even though the plugin only
ever used one, and a large commit had to be split with a `batch_id = 0`
placeholder dance plus a trailing empty-ops call to release the batch.

Replace this with a single in-progress sync, identified by a session id:

- `create_batch` -> `start_sync()`: takes no args; returns a
  `StartSyncResult` (`Started { session_id }` or `Busy { owner,
  idle_for_secs }`). At most one sync runs at a time. The owner can
  reclaim their own sync immediately (retrying a failed deploy); a
  different caller may reclaim it only after it goes stale (30s idle),
  otherwise gets `Busy` so a teammate can't barge into an active deploy.
- `commit_batch` -> `execute_operations({ session_id, operations,
  is_final })`: every call carries the real session id, which the
  canister validates. The last call sets `is_final` to finalize the sync
  (release the lock, drop staged chunks). This removes the `batch_id = 0`
  placeholder hack and the trailing empty-ops cleanup call.
- `delete_batch` -> `cancel_sync({ session_id })`: owner-only release of
  the lock on abort; abandonment is otherwise handled by staleness.

Also unify the wire types for what are just counters/timestamps:

- BatchId/ChunkId: candid `nat` -> `nat64`, Rust `Nat` -> `u64`.
- Time: candid `int` -> `nat64`, Rust `Int` -> `u64` (timestamps are
  always non-negative).

This drops the Nat/Int conversion helpers (batch_id_to_u64,
timestamp_to_u64), the overflow panics, and the "Nat isn't Ord" test
workaround. The session id counter is persisted across upgrades so ids
are never reused; the in-flight session and its chunks are not persisted.

+++ crates/canister-core/src/asset.rs
@@ -20,7 +20,7 @@ use crate::http::{
     CallbackFunc, HeaderField, HttpResponse, StreamingCallbackToken, StreamingStrategy,
 };
 use crate::rc_bytes::RcBytes;
-use candid::{CandidType, Deserialize, Int, Nat};
+use candid::{CandidType, Deserialize, Nat};
 use ic_certification::Hash;
 use ic_representation_independent_hash::Value;
 use serde_bytes::ByteBuf;
@@ -49,7 +49,8 @@ pub fn encoding_certification_order<'a>(
 
 const STATUS_CODES_TO_CERTIFY: [u16; 2] = [200, 304];
 
-pub(crate) type Timestamp = Int;
+/// Nanoseconds since the Unix epoch (`ic0.time()`); always non-negative.
+pub(crate) type Timestamp = u64;
 
 #[derive(Default, Clone, Debug)]
 pub struct AssetEncoding {
